refactor: api url, 클래스 정리 #58
