+++ server/src/main/java/com/hobbyt/domain/member/controller/AuthController.java
@@ -7,7 +7,6 @@ import javax.servlet.http.HttpServletResponse;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -16,7 +15,6 @@ import org.springframework.web.bind.annotation.RestController;
 
 import com.hobbyt.domain.member.dto.request.EmailRequest;
 import com.hobbyt.domain.member.service.AuthService;
-import com.hobbyt.global.security.member.MemberDetails;
 
 import lombok.RequiredArgsConstructor;
 
@@ -53,17 +51,6 @@ public class AuthController {
 		String refreshToken = request.getHeader(REFRESH_TOKEN_HEADER);
 
 		authService.logout(accessToken, refreshToken);
-
-		return new ResponseEntity(HttpStatus.OK);
-	}
-
-	@PostMapping("/withdrawal")
-	public ResponseEntity withdraw(@AuthenticationPrincipal MemberDetails memberDetails, HttpServletRequest request,
-		HttpServletResponse response) {
-
-		String accessToken = request.getHeader(AUTH_HEADER).substring(7);
-
-		authService.withdraw(accessToken, memberDetails.getUsername());
 
 		return new ResponseEntity(HttpStatus.OK);
 	}
