+++ server/src/main/java/com/hobbyt/domain/member/controller/MemberController.java
@@ -1,7 +1,13 @@
 package com.hobbyt.domain.member.controller;
 
+import static com.hobbyt.global.security.constants.AuthConstants.*;
+
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -11,6 +17,7 @@ import org.springframework.web.bind.annotation.RestController;
 
 import com.hobbyt.domain.member.dto.request.SignupRequest;
 import com.hobbyt.domain.member.service.MemberService;
+import com.hobbyt.global.security.member.MemberDetails;
 
 import lombok.RequiredArgsConstructor;
 
@@ -32,5 +39,16 @@ public class MemberController {
 	public ResponseEntity signup(@Validated @RequestBody SignupRequest signupRequest) {
 		Long id = memberService.createUser(signupRequest);
 		return new ResponseEntity(id, HttpStatus.CREATED);
+	}
+
+	@PostMapping("/myPage/delete")
+	public ResponseEntity withdraw(@AuthenticationPrincipal MemberDetails memberDetails, HttpServletRequest request,
+		HttpServletResponse response) {
+
+		String accessToken = request.getHeader(AUTH_HEADER).substring(7);
+
+		memberService.withdraw(accessToken, memberDetails.getUsername());
+
+		return new ResponseEntity(HttpStatus.OK);
 	}
 }
