+++ server/src/main/java/com/hobbyt/domain/member/service/MemberService.java
@@ -1,5 +1,7 @@
 package com.hobbyt.domain.member.service;
 
+import static com.hobbyt.global.security.constants.AuthConstants.*;
+
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -8,6 +10,9 @@ import com.hobbyt.domain.member.dto.request.SignupRequest;
 import com.hobbyt.domain.member.entity.Member;
 import com.hobbyt.domain.member.repository.MemberRepository;
 import com.hobbyt.global.error.exception.MemberExistException;
+import com.hobbyt.global.error.exception.MemberNotExistException;
+import com.hobbyt.global.redis.RedisService;
+import com.hobbyt.global.security.jwt.JwtTokenProvider;
 
 import lombok.RequiredArgsConstructor;
 
@@ -17,6 +22,8 @@ import lombok.RequiredArgsConstructor;
 public class MemberService {
 	private final MemberRepository memberRepository;
 	private final PasswordEncoder passwordEncoder;
+	private final JwtTokenProvider jwtTokenProvider;
+	private final RedisService redisService;
 
 	@Transactional
 	public Long createUser(SignupRequest signupRequest) {
@@ -31,5 +38,23 @@ public class MemberService {
 		if (memberRepository.existsByEmail(email)) {
 			throw new MemberExistException();
 		}
+	}
+
+	@Transactional
+	public void withdraw(final String accessToken, final String email) {
+		Long expiration = jwtTokenProvider.calculateExpiration(accessToken);
+
+		redisService.deleteValue(email);
+
+		if (expiration > 0) {
+			redisService.setValue(accessToken, BLACK_LIST, expiration);
+		}
+
+		Member member = findMemberByEmail(email);
+		member.withdraw();
+	}
+
+	private Member findMemberByEmail(final String email) {
+		return memberRepository.findByEmail(email).orElseThrow(MemberNotExistException::new);
 	}
 }
