+++ server/src/main/java/com/hobbyt/global/config/SecurityConfig.java
@@ -65,9 +65,9 @@ public class SecurityConfig {
 			.and()
 
 			.authorizeRequests()
-			.antMatchers("/api/healthcheck", "/api/auth/**", "/api/members/signup").permitAll()
+			.antMatchers("/api/healthcheck", "/api/auth/code", "/api/members/signup").permitAll()
 			.anyRequest().authenticated();
-
+		
 		return http.build();
 	}
