+++ server/src/test/java/com/hobbyt/domain/member/service/MemberServiceTest.java
@@ -1,8 +1,11 @@
 package com.hobbyt.domain.member.service;
 
+import static com.hobbyt.global.security.constants.AuthConstants.*;
 import static com.hobbyt.util.TestUtil.*;
 import static org.assertj.core.api.Assertions.*;
 import static org.mockito.BDDMockito.*;
+
+import java.util.Optional;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -14,8 +17,11 @@ import org.springframework.security.crypto.password.PasswordEncoder;
 
 import com.hobbyt.domain.member.dto.request.SignupRequest;
 import com.hobbyt.domain.member.entity.Member;
+import com.hobbyt.domain.member.entity.MemberStatus;
 import com.hobbyt.domain.member.repository.MemberRepository;
 import com.hobbyt.global.error.exception.MemberExistException;
+import com.hobbyt.global.redis.RedisService;
+import com.hobbyt.global.security.jwt.JwtTokenProvider;
 
 @ExtendWith(MockitoExtension.class)
 class MemberServiceTest {
@@ -23,6 +29,10 @@ class MemberServiceTest {
 	private MemberRepository memberRepository;
 	@Mock
 	private PasswordEncoder passwordEncoder;
+	@Mock
+	private JwtTokenProvider jwtTokenProvider;
+	@Mock
+	private RedisService redisService;
 	@InjectMocks
 	private MemberService memberService;
 
@@ -58,5 +68,23 @@ class MemberServiceTest {
 
 		then(memberRepository).should(times(1))
 			.existsByEmail(argThat(email -> email.equals(EMAIL)));
+	}
+
+	@DisplayName("회원 탈퇴")
+	@Test
+	void withdraw() {
+		Member member = dummyMember(1L, NICKNAME, EMAIL, PASSWORD);
+		given(jwtTokenProvider.calculateExpiration(anyString())).willReturn(TIMEOUT);
+		given(memberRepository.findByEmail(anyString())).willReturn(Optional.of(member));
+
+		memberService.withdraw(ACCESS_TOKEN, EMAIL);
+
+		assertThat(member.getStatus()).isEqualTo(MemberStatus.WITHDRAWAL);
+		then(jwtTokenProvider).should(times(1)).calculateExpiration(argThat(jws -> jws.equals(ACCESS_TOKEN)));
+		then(redisService).should(times(1)).deleteValue(argThat(key -> key.equals(EMAIL)));
+		then(redisService).should(times(1))
+			.setValue(argThat(key -> key.equals(ACCESS_TOKEN)), argThat(value -> value.equals(BLACK_LIST)),
+				argThat(timeout -> timeout == TIMEOUT));
+		then(memberRepository).should(times(1)).findByEmail(argThat(email -> email.equals(EMAIL)));
 	}
 }
